Moving platform added
There are still some problems with the platforms though... First, when
riding it, it moves down too fast so the gravity can't keep up, and the
sphere starts bouncing (even when setting the platform's 'bounciness' to
0). Second, there are some odd effects when it moves sideways; the
sphere starts sliding as if there was no or little friction (even when
setting the platform's friction to a high value). These things need to
be fixed.