Added first person view
The first person replaced the previous ability to rotate the camera in
the y-direction while in third person view. First person view is
accessed by pressing the left shift key.

I also found a bug: The white spheres that are created when pressing the
button are drawn in front of the text boxes in the debug screen. It
could be because the spheres are after the text boxes in
game._object_list, and are therefore drawn after the text boxes.