Specular light bug fix
Fixed the bug that made the specular highlights be in the wrong place.
Fixed by moving the light source to it's position every frame. (I had
forgotten that the camera is fixed and we move the scene, not the other
way around...)
Also made the light sources visible by drawing a point at their position
when in debug state 2 ("AABB-state").